jext2: corrected indenting

+++ src/jext2/BlockAccess.java
@@ -19,7 +19,7 @@ public class BlockAccess {
 	private static BlockAccess instance;
 
 	/** number of pointers in indirection block */
-    private int ptrs;
+	private int ptrs;
 
 	public BlockAccess(FileChannel blockdev) {
 		if (BlockAccess.instance != null) {
@@ -33,14 +33,14 @@ public class BlockAccess {
 	public ByteBuffer read(long nr) throws IoError {
 		ByteBuffer buf = ByteBuffer.allocate(blocksize);
 		try {
-		    buf.order(ByteOrder.BIG_ENDIAN);
+			buf.order(ByteOrder.BIG_ENDIAN);
 
-		    synchronizer.readLock(nr);
-		    blockdev.read(buf,(((long)(nr & 0xffffffff)) * blocksize));
+			synchronizer.readLock(nr);
+			blockdev.read(buf,((nr & 0xffffffff) * blocksize));
 			synchronizer.readUnlock(nr);
 		} catch (IOException e) {
 			synchronizer.readUnlock(nr);
-		    throw new IoError();
+			throw new IoError();
 		}
 
 		return buf;
@@ -56,22 +56,22 @@ public class BlockAccess {
 	 */
 	@NotThreadSafe(useLock=true)
 	public void readToBufferUnsynchronized(long position, ByteBuffer buf) throws IoError {
-	    try {
-	        buf.order(ByteOrder.BIG_ENDIAN);
-	        blockdev.read(buf, position);
-	    } catch (IOException e) {
-	        throw new IoError();
-	    }
+		try {
+			buf.order(ByteOrder.BIG_ENDIAN);
+			blockdev.read(buf, position);
+		} catch (IOException e) {
+			throw new IoError();
+		}
 	}
 
 	@NotThreadSafe(useLock=true)
 	public void writeFromBufferUnsynchronized(long position, ByteBuffer buf) throws IoError {
-	    buf.order(ByteOrder.BIG_ENDIAN);
-	    try {
-            blockdev.write(buf, position);
-        } catch (IOException e) {
-            throw new IoError();
-        }
+		buf.order(ByteOrder.BIG_ENDIAN);
+		try {
+			blockdev.write(buf, position);
+		} catch (IOException e) {
+			throw new IoError();
+		}
 	}
 
 	/**
@@ -79,48 +79,48 @@ public class BlockAccess {
 	 * @throws IOException
 	 */
 	public void zeroOut(long nr, int start, int end) throws IoError {
-	    ByteBuffer zeros = ByteBuffer.allocate((end-start)+1);
-	    writePartial(nr, start, zeros);
+		ByteBuffer zeros = ByteBuffer.allocate((end-start)+1);
+		writePartial(nr, start, zeros);
 	}
 
 	/** Write a whole block to the logical address nr on disk */
 	public void write(long nr, ByteBuffer buf) throws IoError {
 		buf.rewind();
 		try {
-		    blockdev.position(((long)(nr & 0xffffffff)) * blocksize);
+			blockdev.position((nr & 0xffffffff) * blocksize);
 
-		    synchronizer.writeLock(nr);
-		    blockdev.write(buf);
-		    synchronizer.writeUnlock(nr);
+			synchronizer.writeLock(nr);
+			blockdev.write(buf);
+			synchronizer.writeUnlock(nr);
 		} catch (IOException e) {
-		    synchronizer.writeUnlock(nr);
-		    throw new IoError();
+			synchronizer.writeUnlock(nr);
+			throw new IoError();
 		}
 	}
 
-    public void dumpByteBuffer(ByteBuffer buf) {
-	    try {
-		    while (buf.hasRemaining()) {
-			    for (int i=0; i<8; i++) {
-				    System.out.print(buf.get());
-				    System.out.print("\t\t");
-			    }
-			    System.out.println();
-		    }
-	    } catch (Exception e) {
-	    }
+	public void dumpByteBuffer(ByteBuffer buf) {
+		try {
+			while (buf.hasRemaining()) {
+				for (int i=0; i<8; i++) {
+					System.out.print(buf.get());
+					System.out.print("\t\t");
+				}
+				System.out.println();
+			}
+		} catch (Exception e) {
+		}
 	}
 
-    /**
-     * Force changes to disc
-     */
-    public void sync() throws IoError {
-        try {
-            blockdev.force(false);
-        } catch (IOException e) {
-            throw new IoError();
-        }
-    }
+	/**
+	 * Force changes to disc
+	 */
+	public void sync() throws IoError {
+		try {
+			blockdev.force(false);
+		} catch (IOException e) {
+			throw new IoError();
+		}
+	}
 
 
 	/**
@@ -129,117 +129,117 @@ public class BlockAccess {
 	 * @throws IOException
 	 */
 	public void writePartial(long nr, long offset, ByteBuffer buf) throws IoError {
-        if (offset + buf.limit() > blocksize)
-            throw new IllegalArgumentException("attempt to write over block boundries" + buf + ", " + offset);
+		if (offset + buf.limit() > blocksize)
+			throw new IllegalArgumentException("attempt to write over block boundries" + buf + ", " + offset);
 
-        try {
-            buf.rewind();
-		    synchronizer.writeLock(nr);
-            blockdev.write(buf, ((((long)(nr & 0xffffffff)) * blocksize) + offset));
-		    synchronizer.writeUnlock(nr);
+		try {
+			buf.rewind();
+			synchronizer.writeLock(nr);
+			blockdev.write(buf, (((nr & 0xffffffff) * blocksize) + offset));
+			synchronizer.writeUnlock(nr);
 
-        } catch (IOException e) {
-		    synchronizer.writeUnlock(nr);
-            throw new IoError();
-        }
+		} catch (IOException e) {
+			synchronizer.writeUnlock(nr);
+			throw new IoError();
+		}
 	}
 
 	public void initialize(Superblock superblock) {
-	    blocksize = superblock.getBlocksize();
+		blocksize = superblock.getBlocksize();
 		ptrs = superblock.getAddressesPerBlock();
 	}
 
 	/**
-     * Read block pointers from block.
-     * Note: Zero pointers are not skipped
-     * @param   dataBlock   physical block number
-     * @param   start       index of first pointer to retrieve
-     * @param   end         index of last pointer to retrieve
-     * @return  array containing all block numbers in block
-     */
-    public long[] readBlockNrsFromBlock(long dataBlock, int start, int end) throws IoError {
-        int numEntries = (end-start)+1;
-        long[] result = new long[numEntries];
-        ByteBuffer buffer = ByteBuffer.allocate(numEntries*4);
+	 * Read block pointers from block.
+	 * Note: Zero pointers are not skipped
+	 * @param   dataBlock   physical block number
+	 * @param   start       index of first pointer to retrieve
+	 * @param   end         index of last pointer to retrieve
+	 * @return  array containing all block numbers in block
+	 */
+	public long[] readBlockNrsFromBlock(long dataBlock, int start, int end) throws IoError {
+		int numEntries = (end-start)+1;
+		long[] result = new long[numEntries];
+		ByteBuffer buffer = ByteBuffer.allocate(numEntries*4);
 
-	    synchronizer.readLock(dataBlock);
-	    try {
-	    	readToBufferUnsynchronized(dataBlock*blocksize + start*4, buffer);
-	    } catch (IoError e) {
-	    	synchronizer.readUnlock(dataBlock);
-	    	throw e;
-	    }
-    	synchronizer.readUnlock(dataBlock);
+		synchronizer.readLock(dataBlock);
+		try {
+			readToBufferUnsynchronized(dataBlock*blocksize + start*4, buffer);
+		} catch (IoError e) {
+			synchronizer.readUnlock(dataBlock);
+			throw e;
+		}
+		synchronizer.readUnlock(dataBlock);
 
-        for (int i=0; i<numEntries; i++) {
-            result[i] = Ext2fsDataTypes.getLE32U(buffer, i*4);
-        }
+		for (int i=0; i<numEntries; i++) {
+			result[i] = Ext2fsDataTypes.getLE32U(buffer, i*4);
+		}
 
-        return result;
-    }
+		return result;
+	}
 
-    /**
-     * Read block pointers from block. Skip pointers that are zero.
-     * Note: Since we return a list, you cannot use indices computed
-     * for the block with the result.
-     * @param   dataBlock   physical block number
-     * @param   start       index of first pointer to retrieve
-     * @param   end         index of last pointer to retrieve
-     * @return  list of non-zero block numbers found in block
-     */
-    public LinkedList<Long>
-    readBlockNrsFromBlockSkipZeros(long dataBlock, int start, int end) throws IoError {
-        int numEntries = (end-start)+1;
-        LinkedList<Long> result = new LinkedList<Long>();
-        ByteBuffer buffer = ByteBuffer.allocate(numEntries*4);
+	/**
+	 * Read block pointers from block. Skip pointers that are zero.
+	 * Note: Since we return a list, you cannot use indices computed
+	 * for the block with the result.
+	 * @param   dataBlock   physical block number
+	 * @param   start       index of first pointer to retrieve
+	 * @param   end         index of last pointer to retrieve
+	 * @return  list of non-zero block numbers found in block
+	 */
+	public LinkedList<Long>
+	readBlockNrsFromBlockSkipZeros(long dataBlock, int start, int end) throws IoError {
+		int numEntries = (end-start)+1;
+		LinkedList<Long> result = new LinkedList<Long>();
+		ByteBuffer buffer = ByteBuffer.allocate(numEntries*4);
 
-	    synchronizer.readLock(dataBlock);
-	    try {
-	    	readToBufferUnsynchronized(dataBlock*blocksize + start*4, buffer);
-	    } catch (IoError e) {
-	    	synchronizer.readUnlock(dataBlock);
-	    	throw e;
-	    }
-    	synchronizer.readUnlock(dataBlock);
+		synchronizer.readLock(dataBlock);
+		try {
+			readToBufferUnsynchronized(dataBlock*blocksize + start*4, buffer);
+		} catch (IoError e) {
+			synchronizer.readUnlock(dataBlock);
+			throw e;
+		}
+		synchronizer.readUnlock(dataBlock);
 
-        for (int i=0; i<numEntries; i++) {
-            long tmp = Ext2fsDataTypes.getLE32U(buffer, i*4);
-            if (tmp > 0)
-                result.add(tmp);
-        }
+		for (int i=0; i<numEntries; i++) {
+			long tmp = Ext2fsDataTypes.getLE32U(buffer, i*4);
+			if (tmp > 0)
+				result.add(tmp);
+		}
 
-        return result;
-    }
+		return result;
+	}
 
-    /**
-     * Read all block pointers from block.
-     * @param   dataBlock   physical block number
-     * @return  array with all pointers stored in block
-     */
-    public long[] readAllBlockNumbersFromBlock(long dataBlock) throws IoError {
-        return readBlockNrsFromBlock(dataBlock, 0, ptrs-1);
-    }
+	/**
+	 * Read all block pointers from block.
+	 * @param   dataBlock   physical block number
+	 * @return  array with all pointers stored in block
+	 */
+	public long[] readAllBlockNumbersFromBlock(long dataBlock) throws IoError {
+		return readBlockNrsFromBlock(dataBlock, 0, ptrs-1);
+	}
 
-    /**
-     * Read all block pointers from block. Skip Zero pointers
-     * @param   dataBlock   physical block number
-     * @return  list with all non-zero pointers in block
-     */
-    public LinkedList<Long>
-    readAllBlockNumbersFromBlockSkipZero(long dataBlock) throws IoError {
-        return readBlockNrsFromBlockSkipZeros(dataBlock, 0, ptrs-1);
-    }
+	/**
+	 * Read all block pointers from block. Skip Zero pointers
+	 * @param   dataBlock   physical block number
+	 * @return  list with all non-zero pointers in block
+	 */
+	public LinkedList<Long>
+	readAllBlockNumbersFromBlockSkipZero(long dataBlock) throws IoError {
+		return readBlockNrsFromBlockSkipZeros(dataBlock, 0, ptrs-1);
+	}
 
-    /**
-     * Read single block pointer from block.
-     * @param   dataBlock   physical block number
-     * @param   index       index of block number to retrieve
-     */
-    public long readBlockNumberFromBlock(long dataBlock, int index) throws IoError {
-        return (readBlockNrsFromBlock(dataBlock, index, index))[0];
-    }
+	/**
+	 * Read single block pointer from block.
+	 * @param   dataBlock   physical block number
+	 * @param   index       index of block number to retrieve
+	 */
+	public long readBlockNumberFromBlock(long dataBlock, int index) throws IoError {
+		return (readBlockNrsFromBlock(dataBlock, index, index))[0];
+	}
 
-    public static BlockAccess getInstance() {
+	public static BlockAccess getInstance() {
 		return BlockAccess.instance;
 	}
 }
